client/pyproject.toml
[project]
name = "guardlens-client"
version = "0.1.0"
description = "GuardianLens client — captures screenshots and streams them to the GuardianLens server."
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "GuardianLens Team" }]

dependencies = [
    "mss>=9.0.1",
    "Pillow>=10.0.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
]

[project.scripts]
guardlens-client = "guardlens_client.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/guardlens_client"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.0",
    "pytest>=8.3.0",
]
